refactor(active-memory): split runtime modules
